Update vnet gateway properties

+++ src/ResourceManager/Network/Commands.Network/LocalNetworkGateway/NewAzureLocalNetworkGatewayCommand.cs
@@ -66,7 +66,7 @@ namespace Microsoft.Azure.Commands.Network
             Mandatory = false,
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "The local network gateway's ASN")]
-        public long Asn { get; set; }
+        public uint Asn { get; set; }
 
         [Parameter(
             Mandatory = false,
